backend/requirements.txt
# FastAPI and Server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.9
aiofiles==24.1.0
gunicorn==23.0.0

# Machine Learning
tensorflow==2.17.0
numpy>=1.23.5,<2.0.0
pillow==11.0.0
opencv-python==4.10.0.84

# MongoDB
pymongo==4.9.2
dnspython==2.7.0

# Firebase
firebase-admin==6.5.0

# Authentication
python-jose[cryptography]==3.3.0
PyJWT==2.9.0
cryptography==43.0.1
passlib[bcrypt]==1.7.4

# Utilities
python-dotenv==1.0.1
requests==2.32.3
